+++ src/lib/pipeline/BalancedPortAllocator.cc
@@ -38,7 +38,7 @@ void BalancedPortAllocator::issued(uint16_t port) {
 void BalancedPortAllocator::deallocate(uint16_t port) { issued(port); }
 
 void BalancedPortAllocator::setRSSizeGetter(
-    std::function<void(std::vector<uint64_t>&)> rsSizes) {
+    std::function<void(std::vector<uint32_t>&)> rsSizes) {
   rsSizes_ = rsSizes;
 }
